requirements.txt
flask
requests
instagrapi
python-dotenv
pillow 